+++ fxmanifest.lua
@@ -4,7 +4,7 @@ lua54 'yes'
 
 author 'oRit_907'
 description 'A simple script to show who is currently talking for PMA-Voice.'
-version '2.1.0'
+version '2.2.0'
 
 repository 'https://github.com/oRit-907/ActiveSpeaker'